requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.3
PyQt5==5.15.11
PyQt5_sip==12.16.1
python-dotenv==1.0.1
Requests==2.32.3
